Add Multiversion compatiblity
This commit returns the 1.2 Mod-Files to the 1.2 Folder.
The Textures and Translation are put into Common, as it should be common.
The Assemblies, Defs, Patches, References and Source are by Version, since they are very likely incompatible between RimWorld versions and need Version specific implementations.

# This is not cross-compatible!

This works with either RW 1.2 or RW 1.3, but using a RW 1.2 Save and loading it in RW 1.3 results in undefined behaviour and might not work at all. Always start a new Save in a new RimWorld Major Version.